backend/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
gunicorn==23.0.0
whitenoise==6.7.0
django-cors-headers==4.4.0
psycopg2-binary==2.9.10
dj-database-url==2.3.0
python-dotenv==1.1.1

# (Opcionales tuyos, si los usas)
google-generativeai==0.8.5
google-api-python-client==2.183.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.71.2
httplib2==0.31.0
requests==2.32.5
tqdm==4.67.1
cachetools==5.5.2
protobuf==5.29.5
proto-plus==1.26.1
uritemplate==4.2.0
typing-extensions==4.15.0
tzdata==2025.2
sqlparse==0.5.3
charset_normalizer==3.4.3
idna==3.10
urllib3==2.5.0
colorama==0.4.6
uvicorn[standard]==0.30.6
typing_extensions==4.15.0
